cmd: Move shared code in the cmd into cmd/common package.
