+++ cmd/common/client_builder.go
@@ -1,0 +1,59 @@
+package common
+
+import (
+	"github.com/golang/glog"
+	securityclientset "github.com/openshift/client-go/security/clientset/versioned"
+	mcfgclientset "github.com/openshift/machine-config-operator/pkg/generated/clientset/versioned"
+	apiext "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+)
+
+// ClientBuilder can create a variety of kubernetes client interface
+// with its embeded rest.Config.
+type ClientBuilder struct {
+	config *rest.Config
+}
+
+// ClientOrDie returns the kubernetes client interface for machine config.
+func (cb *ClientBuilder) MachineConfigClientOrDie(name string) mcfgclientset.Interface {
+	return mcfgclientset.NewForConfigOrDie(rest.AddUserAgent(cb.config, name))
+}
+
+// ClientOrDie returns the kubernetes client interface for general kubernetes objects.
+func (cb *ClientBuilder) KubeClientOrDie(name string) kubernetes.Interface {
+	return kubernetes.NewForConfigOrDie(rest.AddUserAgent(cb.config, name))
+}
+
+// ClientOrDie returns the kubernetes client interface for security related kubernetes objects
+// such as pod security policy, security context.
+func (cb *ClientBuilder) SecurityClientOrDie(name string) securityclientset.Interface {
+	return securityclientset.NewForConfigOrDie(rest.AddUserAgent(cb.config, name))
+}
+
+// ClientOrDie returns the kubernetes client interface for extended kubernetes objects.
+func (cb *ClientBuilder) APIExtClientOrDie(name string) apiext.Interface {
+	return apiext.NewForConfigOrDie(rest.AddUserAgent(cb.config, name))
+}
+
+// NewClientBuilder returns a *ClientBuilder with the given kubeconfig.
+func NewClientBuilder(kubeconfig string) (*ClientBuilder, error) {
+	var config *rest.Config
+	var err error
+
+	if kubeconfig != "" {
+		glog.V(4).Infof("Loading kube client config from path %q", kubeconfig)
+		config, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
+	} else {
+		glog.V(4).Infof("Using in-cluster kube client config")
+		config, err = rest.InClusterConfig()
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	return &ClientBuilder{
+		config: config,
+	}, nil
+}
