+++ cmd/common/doc.go
@@ -1,0 +1,3 @@
+// Package common defines the types and functions that are shared by
+// multiple commands in the cmd package.
+package common
